lactation finalized. May merge it into genitals.cs instead of splitting it as a partial class, but it was easier to develop in a (relative) valuum. creatures updated so stats that are universal to all creatures (lust, libido, sensitivity, and corruption) are now in the right spot. increase, decrease and set functions implemented for each one.
Updated code to use a Guid (Unique identifier) instead of integrating the creature directly. allows modular code with an empty Guid, which signifies it's not attached to any creature. each item updated to use a fallback if the default is unavailable. will likely need more defaults as we go forward.