Massive overhaul to how the display works - now, everything should either be passed a display or asked to create one. From there, you can either build off an existing one, if applicable, or completely ignore it if that's your desire. In essence, this is the same as the old code, but the old code hid this away and forced some crazy work-arounds if you actually needed any control over how pages display.
Reworked everything that dealt with displays to work with the new format (and i mean everything, holy shit). Moved a ton of files around so the namespaces made more sense. Note that this fix has not been applied to various encounters and the like; which were not implemented as it was. They will be fully implemented in later updates.

Updated the GUI to work with this new format; which required very little tweaking - one nice bonus of separation of concerns, it seems. Will likely need more minor reworks to handle toast notifications, though that actually needs to be supported in the framework first.

Created new helper to deal with inventory/adding items. This has variations for events, which simply add the item then return execution to the event parser (time engine) and for standard gameplay, which uses a callback with the current display passed along.