Complete and hopefully final overhaul to the system of body parts. Added a read-only option to basically everything body part related, so you can still reference the data and not have to worry about people messing with it. originally, this was a means to allow data classes to be reused for various text, but it expanded into its own thing. honestly, though, i'd be surprised if they ever get heavy use.
Data classes are now optional; they are used as data change classes so you can be lazy and dump all the data in a class instead of manually checking everything and creating a custom event for it.

Likely will require some additional class be added to these types for serialization and deserialization, but that's a later worry.

Restructured code to make it clear text describing changes should occur BEFORE the change occurs, not afterward - it's much too difficult to correctly copy ALL the data needed for this automatically (anything that's lazy will break this, for example, unless you know what you're doing), and far too easy to mess up. That said, it's really easy to copy data manually, and just parse it as needed, so this is the desired behavior when you really want to customize it.

Temporarily removed transformations and their items because they assumed data afterward, not before. Will be added back when they are corrected.

Reorganized some classes so the data was easier to follow and understand; related functions and members are grouped, and #region preprocessor directives are used.