Converted all the body parts to use the Data Class for the longDescription, which allows them to be used for NPCs and such, and previous types for change and restore type strings. Will likely require another pass for any variations of long/short with various extra options, which i definitely missed.
Added ReadOnly option for piercings, and included them in the data classes for body parts that use them. All body parts should now store all the required data to deserialize them in the data class, with exception to statistics, which i haven't figured out how to deal with yet.

Added a to current data function to the data class, allowing legacy data classes to exist when the other data changes. should in theory make deserialization easier, because programming magic.

Converted all line endings to CR-LF, because windows. Honestly, it doesn't matter which type they use, but they need to be consistent.