implemeted tattoos in all classes that will use them. made sure piercings and tattoos are cleaned up. updated genitals to have everything public in the base class in the data class. moved the currently implemented genitals texts to a common static class with an interface allowing them to be used in either the data class or the source class.
Temporarily gave up on implementing remaining body part strings (too tedious), will leave them for later. currently reimplementing the various items; cleaned up and added back the previous transformation based items. current plan is to create them and simply not implement the text - leave it not implemented and handle the text later after the code is finished.
Commit  may be in a breaking state, not sure. this will be quickly fixed, but it's simpler to commit this, checkout the other branch, then port what i need to than it is to stash the changes and go that route.