Created a bunch of transformations and items that use them. very few, if any, have any text or even any calls to text functions - they will be added in post after all of the hard work (aka porting the code) is done.
Reworked Nipples and Breasts - they now are a little less complicated, though they still have some unneeded complexity - i'll see what i can do balancing complexity and coupling. Nipples now have a single aggregate class that provides the behavior for all of them - this is what the original code did, just far more conviluted. this is still conviluted, of course, but much less so. all classes that referenced nipples now updated accordingly.

Removed the genital perk helpers - they are no longer necessary because perk data is stored in the genitals and the related parts are only created after the perks have been initialized.

Updated several classes to use abstract functions instead of delegates - it's generally easier to deal with funky cases that require local variables this way, and it's a bit easier to understand to boot. not all of the classes do this, just the ones that are implemented in the frontend. I'll likely need to do this for a few more classes as well.