updated frontend perks to use modern modifiier paradigm. they will now compile correctoy. updated extended perk modifiers to use this same paradigm.
removed combat loss flag from capacity items. updated existing code accordingly. this is likely not the final change; we will like need to pass a combat state object to these items so they can manipulate it accordingly - maybe an item causes the attacker/defender to flee, or stuns someone, etc.

adjusted frontend to remove the 'f' float designation on various numerics, as we're using doubles now. added '.0' where necessary to facilitate that.

This commit is a compile-safe commit - the code can be run after this commit.